requirements.txt
kagglehub
fastapi
uvicorn
tensorflow
azure.storage.blob
python-multipart